requirements.txt
pandas==2.1.4
google-generativeai==0.8.3
python-pptx==0.6.23
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
streamlit==1.29.0
pymongo==4.6.0
sqlalchemy==2.0.25
python-dotenv==1.0.0